requirements.txt
pandas
matplotlib
numpy
seaborn
scikit-learn
imblearn
xgboost
kafka-python
requests
flask